requirements.txt
kiteconnect
python-dotenv
pandas
openpyxl
numpy
matplotlib
seaborn
